Create sharepoint connector object

Create a new sharepoint connector object. See ConnectorSharepoint for details.

Usage

connector_sharepoint(
  site_url,
  token = get_token(),
  folder = "",
  ...,
  extra_class = NULL
)

Arguments

site_url

The URL of the Sharepoint site

token

The Azure token. By default, it will be retrieve by get_token

folder

The path of the folder to interact with, if you don't want to interact with the root folder "Documents" #nolint

...

Additional parameters to pass to the ConnectorSharepoint object

extra_class

character Extra class added to the object.

Value

A new ConnectorSharepoint object

Details

The extra_class parameter allows you to create a subclass of the ConnectorSharepoint object. This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods, while still inheriting the methods from the ConnectorSharepoint object.

Examples

if (FALSE) {
# Connect
cs <- connector_sharepoint(Sys.getenv("SHAREPOINT_SITE_URL"))
cs

# Create subclass connection
cs_subclass <- connector_sharepoint(Sys.getenv("SHAREPOINT_SITE_URL"),
  extra_class = "subclass"
)

cs_subclass
class(cs_subclass)
}